CHEAT_007_d add connectedUserId + setter and getter in BluetoothProvider; make unit tests run; change data type of id in database to long [PK,SE]

+++ app/src/main/java/at/tugraz/ist/swe/cheatapp/Message.java
@@ -12,11 +12,11 @@ public class Message {
     @PrimaryKey(autoGenerate = true)
     private int messageId;
 
-    private int userId;
+    private long userId;
     private String messageText;
     private boolean messageSent;
 
-    public Message(int userId, String messageText, boolean messageSent) {
+    public Message(long userId, String messageText, boolean messageSent) {
         this.userId = userId;
         this.messageText = messageText;
         this.messageSent = messageSent;
@@ -24,7 +24,7 @@ public class Message {
 
     public Message(String jsonMessageString, boolean messageSent) throws JSONException {
         JSONObject jsonMessage = new JSONObject(jsonMessageString);
-        this.userId = jsonMessage.getInt("userId");
+        this.userId = jsonMessage.getLong("userId");
         this.messageText = jsonMessage.getString("messageText");
         this.messageSent = messageSent;
     }
@@ -37,11 +37,11 @@ public class Message {
         this.messageId = messageId;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
